Check API Keys for Various Web of Science APIs

Check API Keys for Various Web of Science APIs

ws_construct_api_key(api_key = NULL, error = TRUE, key_option, sys_env,
  key_filename_option, api_name = "")

ws_api_key(api_key = NULL, error = TRUE, api = names(ws_list_apis()))

ws_api_key_header(api_key = NULL)

ws_list_apis()

ws_auth_header(api_key = NULL)

Arguments

api_key
Web of Science API key

error
Should the function error if api_key = NULL after looking in various places? (See notes)

key_option
option to grab API key using getOption

sys_env
System environment variable to grab API key from Sys.getenv

key_filename_option
Option retrieved from getOption, which is a location of a file with the API key.

api_name
Name of API key

api
an endpoint for the API, see ws_list_apis

Value

API key character string

Note

You can either set the API key using option(ws_API_key) or have it accessible by api_key = Sys.getenv('API_KEY'), where API is one of the Web of Science APIs, such as INCITES

Examples

api_key = NULL
error = FALSE
  api_key = ws_construct_api_key(
  api_key = api_key,
  error = error,
  key_option = "ws_incites_key",
  sys_env = "INCITES_KEY",
  key_filename_option = "ws_incites_key_filename",
  api_name = "Incites")
ws_list_apis()

#> $incites
#> $incites$key_option
#> [1] "ws_incites_key"
#> 
#> $incites$sys_env
#> [1] "INCITES_KEY"
#> 
#> $incites$key_filename_option
#> [1] "ws_incites_key_filename"
#> 
#> $incites$api_name
#> [1] "Incites"
#> 
#> $incites$endpoint
#> [1] "incites"
#> 
#> $incites$url
#> [1] "https://api.clarivate.com/api/incites"
#> 
#> 
#> $endnote
#> $endnote$key_option
#> [1] "ws_endnote_key"
#> 
#> $endnote$sys_env
#> [1] "ENDNOTE_KEY"
#> 
#> $endnote$key_filename_option
#> [1] "ws_endnote_key_filename"
#> 
#> $endnote$api_name
#> [1] "EndNote"
#> 
#> $endnote$endpoint
#> [1] "endnote"
#> 
#> $endnote$url
#> [1] "https://api.clarivate.com/api/endnote"
#> 
#> 
#> $wos
#> $wos$key_option
#> [1] "ws_wos_key"
#> 
#> $wos$sys_env
#> [1] "WOS_KEY"
#> 
#> $wos$key_filename_option
#> [1] "ws_wos_key_filename"
#> 
#> $wos$api_name
#> [1] "WOS"
#> 
#> $wos$endpoint
#> [1] "wos"
#> 
#> $wos$url
#> [1] "https://api.clarivate.com/api/wos"
#> 
#> 
#> $woslite
#> $woslite$key_option
#> [1] "ws_woslite_key"
#> 
#> $woslite$sys_env
#> [1] "WOSLITE_KEY"
#> 
#> $woslite$key_filename_option
#> [1] "ws_woslite_key_filename"
#> 
#> $woslite$api_name
#> [1] "WOS Lite"
#> 
#> $woslite$endpoint
#> [1] "woslite"
#> 
#> $woslite$url
#> [1] "https://api.clarivate.com/api/woslite"
#> 
#> 
#> $tipms
#> $tipms$key_option
#> [1] "ws_tipms_key"
#> 
#> $tipms$sys_env
#> [1] "TIPMS_KEY"
#> 
#> $tipms$key_filename_option
#> [1] "ws_tipms_key_filename"
#> 
#> $tipms$api_name
#> [1] "IP Management System"
#> 
#> $tipms$endpoint
#> [1] "tipms"
#> 
#> $tipms$url
#> [1] "https://api.clarivate.com/api/tipms"
#> 
#> 

ws_auth_header(api_key = "my_api_key")

#> <request>
#> Headers:
#> * X-API-Key: my_api_key