Web of Science Endpoints

Web of Science Endpoints

ws_endnote_endpoints()

ws_incites_endpoints(out_type = c("json", "xml"))

.ws_incites_endpoints()

ws_tipms_endpoints()

ws_wos_endpoints()

ws_woslite_endpoints()

ws_endpoints(out_type = c("json", "xml"), api = NULL)

Arguments

out_type
output type, either json or json

api
The API being used

Value

A list of endpoints for each API

Examples

ws_incites_endpoints()

#> $DocumentLevelMetricsByInstitutionId
#> $DocumentLevelMetricsByInstitutionId$description
#> [1] "Get metrics by institution Id"
#> 
#> $DocumentLevelMetricsByInstitutionId$verb
#> [1] "GET"
#> 
#> $DocumentLevelMetricsByInstitutionId$mime_type
#> [1] "json"
#> 
#> $DocumentLevelMetricsByInstitutionId$endpoint
#> [1] "/DocumentLevelMetricsByInstitutionId/json"
#> 
#> 
#> $DocumentLevelMetricsByInstitutionIdRecordCount
#> $DocumentLevelMetricsByInstitutionIdRecordCount$description
#> [1] "Get metrics count by year"
#> 
#> $DocumentLevelMetricsByInstitutionIdRecordCount$verb
#> [1] "GET"
#> 
#> $DocumentLevelMetricsByInstitutionIdRecordCount$mime_type
#> [1] "json"
#> 
#> $DocumentLevelMetricsByInstitutionIdRecordCount$endpoint
#> [1] "/DocumentLevelMetricsByInstitutionIdRecordCount/json"
#> 
#> 
#> $DocumentLevelMetricsByUT
#> $DocumentLevelMetricsByUT$description
#> [1] "Get metrics by UT"
#> 
#> $DocumentLevelMetricsByUT$verb
#> [1] "GET"
#> 
#> $DocumentLevelMetricsByUT$mime_type
#> [1] "json"
#> 
#> $DocumentLevelMetricsByUT$endpoint
#> [1] "/DocumentLevelMetricsByUT/json"
#> 
#> 
#> $CheckValidUTs
#> $CheckValidUTs$description
#> [1] "Get invalid UTs"
#> 
#> $CheckValidUTs$verb
#> [1] "GET"
#> 
#> $CheckValidUTs$mime_type
#> [1] "json"
#> 
#> $CheckValidUTs$endpoint
#> [1] "/CheckValidUTs/json"
#> 
#> 
#> $InCitesLastUpdated
#> $InCitesLastUpdated$description
#> [1] "Get last updated information"
#> 
#> $InCitesLastUpdated$verb
#> [1] "GET"
#> 
#> $InCitesLastUpdated$mime_type
#> [1] "json"
#> 
#> $InCitesLastUpdated$endpoint
#> [1] "/InCitesLastUpdated/json"
#> 
#> 

ws_endpoints()

#> $incites
#> $incites$DocumentLevelMetricsByInstitutionId
#> $incites$DocumentLevelMetricsByInstitutionId$description
#> [1] "Get metrics by institution Id"
#> 
#> $incites$DocumentLevelMetricsByInstitutionId$verb
#> [1] "GET"
#> 
#> $incites$DocumentLevelMetricsByInstitutionId$mime_type
#> [1] "json"
#> 
#> $incites$DocumentLevelMetricsByInstitutionId$endpoint
#> [1] "/DocumentLevelMetricsByInstitutionId/json"
#> 
#> 
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount$description
#> [1] "Get metrics count by year"
#> 
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount$verb
#> [1] "GET"
#> 
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount$mime_type
#> [1] "json"
#> 
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount$endpoint
#> [1] "/DocumentLevelMetricsByInstitutionIdRecordCount/json"
#> 
#> 
#> $incites$DocumentLevelMetricsByUT
#> $incites$DocumentLevelMetricsByUT$description
#> [1] "Get metrics by UT"
#> 
#> $incites$DocumentLevelMetricsByUT$verb
#> [1] "GET"
#> 
#> $incites$DocumentLevelMetricsByUT$mime_type
#> [1] "json"
#> 
#> $incites$DocumentLevelMetricsByUT$endpoint
#> [1] "/DocumentLevelMetricsByUT/json"
#> 
#> 
#> $incites$CheckValidUTs
#> $incites$CheckValidUTs$description
#> [1] "Get invalid UTs"
#> 
#> $incites$CheckValidUTs$verb
#> [1] "GET"
#> 
#> $incites$CheckValidUTs$mime_type
#> [1] "json"
#> 
#> $incites$CheckValidUTs$endpoint
#> [1] "/CheckValidUTs/json"
#> 
#> 
#> $incites$InCitesLastUpdated
#> $incites$InCitesLastUpdated$description
#> [1] "Get last updated information"
#> 
#> $incites$InCitesLastUpdated$verb
#> [1] "GET"
#> 
#> $incites$InCitesLastUpdated$mime_type
#> [1] "json"
#> 
#> $incites$InCitesLastUpdated$endpoint
#> [1] "/InCitesLastUpdated/json"
#> 
#> 
#> 
#> $endpoint
#> $endpoint$user_space_attachment_size
#> $endpoint$user_space_attachment_size$verb
#> [1] "GET"
#> 
#> $endpoint$user_space_attachment_size$endpoint
#> [1] "/user_space/attachment_size"
#> 
#> $endpoint$user_space_attachment_size$description
#> [1] "Retrieve the current total space of the all user's attachments"
#> 
#> 
#> $endpoint$user_space_attachments
#> $endpoint$user_space_attachments$verb
#> [1] "GET"
#> 
#> $endpoint$user_space_attachments$endpoint
#> [1] "/user_space/attachments"
#> 
#> $endpoint$user_space_attachments$description
#> [1] "Retrieve the total space available for a user"
#> 
#> 
#> $endpoint$reference_types
#> $endpoint$reference_types$verb
#> [1] "GET"
#> 
#> $endpoint$reference_types$endpoint
#> [1] "/reference/types"
#> 
#> $endpoint$reference_types$description
#> [1] "Retrieves the full collection of all reference types"
#> 
#> 
#> $endpoint$user_space
#> $endpoint$user_space$verb
#> [1] "GET"
#> 
#> $endpoint$user_space$endpoint
#> [1] "/user_space"
#> 
#> $endpoint$user_space$description
#> [1] "Retrieve user's space of the library"
#> 
#> 
#> $endpoint$user_space_references
#> $endpoint$user_space_references$verb
#> [1] "GET"
#> 
#> $endpoint$user_space_references$endpoint
#> [1] "/user_space/references"
#> 
#> $endpoint$user_space_references$description
#> [1] "Retrieve the amount of references that a user currently has, the amount of references a user can have in the library, and the amount of references that a user can still add."
#> 
#> 
#> $endpoint$references_guid_file
#> $endpoint$references_guid_file$verb
#> [1] "POST"
#> 
#> $endpoint$references_guid_file$endpoint
#> [1] "/references/{guid}/file"
#> 
#> $endpoint$references_guid_file$description
#> [1] "Create a file attached to this reference"
#> 
#> 
#> $endpoint$reference_get
#> $endpoint$reference_get$verb
#> [1] "GET"
#> 
#> $endpoint$reference_get$endpoint
#> [1] "/reference"
#> 
#> $endpoint$reference_get$description
#> [1] "Retrieve the reference specified in header x-enws-guid"
#> 
#> 
#> $endpoint$reference_put
#> $endpoint$reference_put$verb
#> [1] "PUT"
#> 
#> $endpoint$reference_put$endpoint
#> [1] "/reference"
#> 
#> $endpoint$reference_put$description
#> [1] "Change a set of a reference bibliographic content fields"
#> 
#> 
#> $endpoint$reference_post
#> $endpoint$reference_post$verb
#> [1] "POST"
#> 
#> $endpoint$reference_post$endpoint
#> [1] "/reference"
#> 
#> $endpoint$reference_post$description
#> [1] "Creates a reference from a set of its fields"
#> 
#> 
#> $endpoint$references_guid_get
#> $endpoint$references_guid_get$verb
#> [1] "GET"
#> 
#> $endpoint$references_guid_get$endpoint
#> [1] "/references/{guid}"
#> 
#> $endpoint$references_guid_get$description
#> [1] "Retreive the set of fields of a specific reference"
#> 
#> 
#> $endpoint$references_guid_put
#> $endpoint$references_guid_put$verb
#> [1] "PUT"
#> 
#> $endpoint$references_guid_put$endpoint
#> [1] "/references/{guid}"
#> 
#> $endpoint$references_guid_put$description
#> [1] "Change a set of this reference bibliographic content fields"
#> 
#> 
#> $endpoint$root
#> $endpoint$root$verb
#> [1] "GET"
#> 
#> $endpoint$root$endpoint
#> [1] "/"
#> 
#> $endpoint$root$description
#> [1] "Retrieve the user's Library object."
#> 
#> 
#> 
#> $woslite
#> $woslite$root
#> $woslite$root$verb
#> [1] "GET"
#> 
#> $woslite$root$endpoint
#> [1] "/"
#> 
#> $woslite$root$description
#> [1] "  Submits a user query and returns results"
#> 
#> 
#> $woslite$query
#> $woslite$query$verb
#> [1] "GET"
#> 
#> $woslite$query$endpoint
#> [1] "/query/{queryId}"
#> 
#> $woslite$query$description
#> [1] "Fetch record(s) by query identifier"
#> 
#> 
#> $woslite$id
#> $woslite$id$endpoint
#> [1] "/id/{uniqueId}"
#> 
#> $woslite$id$verb
#> [1] "GET"
#> 
#> $woslite$id$description
#> [1] "Find record(s) by specific id"
#> 
#> 
#> 
#> $wos
#> $wos$id
#> $wos$id$verb
#> [1] "GET"
#> 
#> $wos$id$endpoint
#> [1] "/"
#> 
#> $wos$id$description
#> [1] "Submits a user query and returns results"
#> 
#> 
#> $wos$query
#> $wos$query$verb
#> [1] "GET"
#> 
#> $wos$query$endpoint
#> [1] "/query/{queryId}"
#> 
#> $wos$query$description
#> [1] "Fetch record(s) by query identifier"
#> 
#> 
#> $wos$related
#> $wos$related$verb
#> [1] "GET"
#> 
#> $wos$related$endpoint
#> [1] "/related"
#> 
#> $wos$related$description
#> [1] "Identify related records"
#> 
#> 
#> $wos$citing
#> $wos$citing$verb
#> [1] "GET"
#> 
#> $wos$citing$endpoint
#> [1] "/citing"
#> 
#> $wos$citing$description
#> [1] "Find citing items"
#> 
#> 
#> $wos$references
#> $wos$references$verb
#> [1] "GET"
#> 
#> $wos$references$endpoint
#> [1] "/references"
#> 
#> $wos$references$description
#> [1] "Find cited references by id"
#> 
#> 
#> $wos$id
#> $wos$id$verb
#> [1] "GET"
#> 
#> $wos$id$endpoint
#> [1] "/id/{uniqueId}"
#> 
#> $wos$id$description
#> [1] "Find record(s) by specific id"
#> 
#> 
#> $wos$recordids
#> $wos$recordids$verb
#> [1] "GET"
#> 
#> $wos$recordids$endpoint
#> [1] "/recordids/{queryId}"
#> 
#> $wos$recordids$description
#> [1] "Return IDs of records for a given query by a query identifier"
#> 
#> 
#> 
#> $tipms
#> $tipms$case_record_post
#> $tipms$case_record_post$verb
#> [1] "POST"
#> 
#> $tipms$case_record_post$endpoint
#> [1] "/CaseRecord/{ipType}"
#> 
#> $tipms$case_record_post$description
#> [1] "Posts (Creates) the fields for case record."
#> 
#> 
#> $tipms$case_record_put
#> $tipms$case_record_put$verb
#> [1] "PUT"
#> 
#> $tipms$case_record_put$endpoint
#> [1] "/CaseRecord/{ipType}/{masterId}"
#> 
#> $tipms$case_record_put$description
#> [1] "Puts (Updates) the fields for specified case record."
#> 
#> 
#> $tipms$file
#> $tipms$file$verb
#> [1] "POST"
#> 
#> $tipms$file$endpoint
#> [1] "/File"
#> 
#> $tipms$file$description
#> [1] "Posts selected file."
#> 
#> 
#> $tipms$health
#> $tipms$health$verb
#> [1] "GET"
#> 
#> $tipms$health$endpoint
#> [1] "/Health"
#> 
#> $tipms$health$description
#> [1] "Reflects the state of the application."
#> 
#> 
#> $tipms$patent_filing_forms
#> $tipms$patent_filing_forms$verb
#> [1] "GET"
#> 
#> $tipms$patent_filing_forms$endpoint
#> [1] "/patentfilingforms"
#> 
#> $tipms$patent_filing_forms$description
#> [1] "Gets (Retrieves) the patent filing forms data."
#> 
#> 
#> $tipms$tempfile
#> $tipms$tempfile$verb
#> [1] "POST"
#> 
#> $tipms$tempfile$endpoint
#> [1] "/TempFile"
#> 
#> $tipms$tempfile$description
#> [1] "Posts (Creates) the selected file."
#> 
#> 
#> $tipms$trademark_reporting
#> $tipms$trademark_reporting$verb
#> [1] "GET"
#> 
#> $tipms$trademark_reporting$endpoint
#> [1] "/trademarkreporting"
#> 
#> $tipms$trademark_reporting$description
#> [1] "Search for Trademarks reporting data."
#> 
#> 
#> 

ws_endpoints(out_type = "xml")

#> $incites
#> $incites$DocumentLevelMetricsByInstitutionId
#> $incites$DocumentLevelMetricsByInstitutionId$description
#> [1] "Get metrics by institution Id"
#> 
#> $incites$DocumentLevelMetricsByInstitutionId$verb
#> [1] "GET"
#> 
#> $incites$DocumentLevelMetricsByInstitutionId$mime_type
#> [1] "xml"
#> 
#> $incites$DocumentLevelMetricsByInstitutionId$endpoint
#> [1] "/DocumentLevelMetricsByInstitutionId/xml"
#> 
#> 
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount$description
#> [1] "Get metrics count by year"
#> 
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount$verb
#> [1] "GET"
#> 
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount$mime_type
#> [1] "xml"
#> 
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount$endpoint
#> [1] "/DocumentLevelMetricsByInstitutionIdRecordCount/xml"
#> 
#> 
#> $incites$DocumentLevelMetricsByUT
#> $incites$DocumentLevelMetricsByUT$description
#> [1] "Get metrics by UT"
#> 
#> $incites$DocumentLevelMetricsByUT$verb
#> [1] "GET"
#> 
#> $incites$DocumentLevelMetricsByUT$mime_type
#> [1] "xml"
#> 
#> $incites$DocumentLevelMetricsByUT$endpoint
#> [1] "/DocumentLevelMetricsByUT/xml"
#> 
#> 
#> $incites$CheckValidUTs
#> $incites$CheckValidUTs$description
#> [1] "Get invalid UTs"
#> 
#> $incites$CheckValidUTs$verb
#> [1] "GET"
#> 
#> $incites$CheckValidUTs$mime_type
#> [1] "xml"
#> 
#> $incites$CheckValidUTs$endpoint
#> [1] "/CheckValidUTs/xml"
#> 
#> 
#> $incites$InCitesLastUpdated
#> $incites$InCitesLastUpdated$description
#> [1] "Get last updated information"
#> 
#> $incites$InCitesLastUpdated$verb
#> [1] "GET"
#> 
#> $incites$InCitesLastUpdated$mime_type
#> [1] "xml"
#> 
#> $incites$InCitesLastUpdated$endpoint
#> [1] "/InCitesLastUpdated/xml"
#> 
#> 
#> 
#> $endpoint
#> $endpoint$user_space_attachment_size
#> $endpoint$user_space_attachment_size$verb
#> [1] "GET"
#> 
#> $endpoint$user_space_attachment_size$endpoint
#> [1] "/user_space/attachment_size"
#> 
#> $endpoint$user_space_attachment_size$description
#> [1] "Retrieve the current total space of the all user's attachments"
#> 
#> 
#> $endpoint$user_space_attachments
#> $endpoint$user_space_attachments$verb
#> [1] "GET"
#> 
#> $endpoint$user_space_attachments$endpoint
#> [1] "/user_space/attachments"
#> 
#> $endpoint$user_space_attachments$description
#> [1] "Retrieve the total space available for a user"
#> 
#> 
#> $endpoint$reference_types
#> $endpoint$reference_types$verb
#> [1] "GET"
#> 
#> $endpoint$reference_types$endpoint
#> [1] "/reference/types"
#> 
#> $endpoint$reference_types$description
#> [1] "Retrieves the full collection of all reference types"
#> 
#> 
#> $endpoint$user_space
#> $endpoint$user_space$verb
#> [1] "GET"
#> 
#> $endpoint$user_space$endpoint
#> [1] "/user_space"
#> 
#> $endpoint$user_space$description
#> [1] "Retrieve user's space of the library"
#> 
#> 
#> $endpoint$user_space_references
#> $endpoint$user_space_references$verb
#> [1] "GET"
#> 
#> $endpoint$user_space_references$endpoint
#> [1] "/user_space/references"
#> 
#> $endpoint$user_space_references$description
#> [1] "Retrieve the amount of references that a user currently has, the amount of references a user can have in the library, and the amount of references that a user can still add."
#> 
#> 
#> $endpoint$references_guid_file
#> $endpoint$references_guid_file$verb
#> [1] "POST"
#> 
#> $endpoint$references_guid_file$endpoint
#> [1] "/references/{guid}/file"
#> 
#> $endpoint$references_guid_file$description
#> [1] "Create a file attached to this reference"
#> 
#> 
#> $endpoint$reference_get
#> $endpoint$reference_get$verb
#> [1] "GET"
#> 
#> $endpoint$reference_get$endpoint
#> [1] "/reference"
#> 
#> $endpoint$reference_get$description
#> [1] "Retrieve the reference specified in header x-enws-guid"
#> 
#> 
#> $endpoint$reference_put
#> $endpoint$reference_put$verb
#> [1] "PUT"
#> 
#> $endpoint$reference_put$endpoint
#> [1] "/reference"
#> 
#> $endpoint$reference_put$description
#> [1] "Change a set of a reference bibliographic content fields"
#> 
#> 
#> $endpoint$reference_post
#> $endpoint$reference_post$verb
#> [1] "POST"
#> 
#> $endpoint$reference_post$endpoint
#> [1] "/reference"
#> 
#> $endpoint$reference_post$description
#> [1] "Creates a reference from a set of its fields"
#> 
#> 
#> $endpoint$references_guid_get
#> $endpoint$references_guid_get$verb
#> [1] "GET"
#> 
#> $endpoint$references_guid_get$endpoint
#> [1] "/references/{guid}"
#> 
#> $endpoint$references_guid_get$description
#> [1] "Retreive the set of fields of a specific reference"
#> 
#> 
#> $endpoint$references_guid_put
#> $endpoint$references_guid_put$verb
#> [1] "PUT"
#> 
#> $endpoint$references_guid_put$endpoint
#> [1] "/references/{guid}"
#> 
#> $endpoint$references_guid_put$description
#> [1] "Change a set of this reference bibliographic content fields"
#> 
#> 
#> $endpoint$root
#> $endpoint$root$verb
#> [1] "GET"
#> 
#> $endpoint$root$endpoint
#> [1] "/"
#> 
#> $endpoint$root$description
#> [1] "Retrieve the user's Library object."
#> 
#> 
#> 
#> $woslite
#> $woslite$root
#> $woslite$root$verb
#> [1] "GET"
#> 
#> $woslite$root$endpoint
#> [1] "/"
#> 
#> $woslite$root$description
#> [1] "  Submits a user query and returns results"
#> 
#> 
#> $woslite$query
#> $woslite$query$verb
#> [1] "GET"
#> 
#> $woslite$query$endpoint
#> [1] "/query/{queryId}"
#> 
#> $woslite$query$description
#> [1] "Fetch record(s) by query identifier"
#> 
#> 
#> $woslite$id
#> $woslite$id$endpoint
#> [1] "/id/{uniqueId}"
#> 
#> $woslite$id$verb
#> [1] "GET"
#> 
#> $woslite$id$description
#> [1] "Find record(s) by specific id"
#> 
#> 
#> 
#> $wos
#> $wos$id
#> $wos$id$verb
#> [1] "GET"
#> 
#> $wos$id$endpoint
#> [1] "/"
#> 
#> $wos$id$description
#> [1] "Submits a user query and returns results"
#> 
#> 
#> $wos$query
#> $wos$query$verb
#> [1] "GET"
#> 
#> $wos$query$endpoint
#> [1] "/query/{queryId}"
#> 
#> $wos$query$description
#> [1] "Fetch record(s) by query identifier"
#> 
#> 
#> $wos$related
#> $wos$related$verb
#> [1] "GET"
#> 
#> $wos$related$endpoint
#> [1] "/related"
#> 
#> $wos$related$description
#> [1] "Identify related records"
#> 
#> 
#> $wos$citing
#> $wos$citing$verb
#> [1] "GET"
#> 
#> $wos$citing$endpoint
#> [1] "/citing"
#> 
#> $wos$citing$description
#> [1] "Find citing items"
#> 
#> 
#> $wos$references
#> $wos$references$verb
#> [1] "GET"
#> 
#> $wos$references$endpoint
#> [1] "/references"
#> 
#> $wos$references$description
#> [1] "Find cited references by id"
#> 
#> 
#> $wos$id
#> $wos$id$verb
#> [1] "GET"
#> 
#> $wos$id$endpoint
#> [1] "/id/{uniqueId}"
#> 
#> $wos$id$description
#> [1] "Find record(s) by specific id"
#> 
#> 
#> $wos$recordids
#> $wos$recordids$verb
#> [1] "GET"
#> 
#> $wos$recordids$endpoint
#> [1] "/recordids/{queryId}"
#> 
#> $wos$recordids$description
#> [1] "Return IDs of records for a given query by a query identifier"
#> 
#> 
#> 
#> $tipms
#> $tipms$case_record_post
#> $tipms$case_record_post$verb
#> [1] "POST"
#> 
#> $tipms$case_record_post$endpoint
#> [1] "/CaseRecord/{ipType}"
#> 
#> $tipms$case_record_post$description
#> [1] "Posts (Creates) the fields for case record."
#> 
#> 
#> $tipms$case_record_put
#> $tipms$case_record_put$verb
#> [1] "PUT"
#> 
#> $tipms$case_record_put$endpoint
#> [1] "/CaseRecord/{ipType}/{masterId}"
#> 
#> $tipms$case_record_put$description
#> [1] "Puts (Updates) the fields for specified case record."
#> 
#> 
#> $tipms$file
#> $tipms$file$verb
#> [1] "POST"
#> 
#> $tipms$file$endpoint
#> [1] "/File"
#> 
#> $tipms$file$description
#> [1] "Posts selected file."
#> 
#> 
#> $tipms$health
#> $tipms$health$verb
#> [1] "GET"
#> 
#> $tipms$health$endpoint
#> [1] "/Health"
#> 
#> $tipms$health$description
#> [1] "Reflects the state of the application."
#> 
#> 
#> $tipms$patent_filing_forms
#> $tipms$patent_filing_forms$verb
#> [1] "GET"
#> 
#> $tipms$patent_filing_forms$endpoint
#> [1] "/patentfilingforms"
#> 
#> $tipms$patent_filing_forms$description
#> [1] "Gets (Retrieves) the patent filing forms data."
#> 
#> 
#> $tipms$tempfile
#> $tipms$tempfile$verb
#> [1] "POST"
#> 
#> $tipms$tempfile$endpoint
#> [1] "/TempFile"
#> 
#> $tipms$tempfile$description
#> [1] "Posts (Creates) the selected file."
#> 
#> 
#> $tipms$trademark_reporting
#> $tipms$trademark_reporting$verb
#> [1] "GET"
#> 
#> $tipms$trademark_reporting$endpoint
#> [1] "/trademarkreporting"
#> 
#> $tipms$trademark_reporting$description
#> [1] "Search for Trademarks reporting data."
#> 
#> 
#> 

ws_endpoints(api = "incites")

#> $DocumentLevelMetricsByInstitutionId
#> $DocumentLevelMetricsByInstitutionId$description
#> [1] "Get metrics by institution Id"
#> 
#> $DocumentLevelMetricsByInstitutionId$verb
#> [1] "GET"
#> 
#> $DocumentLevelMetricsByInstitutionId$mime_type
#> [1] "json"
#> 
#> $DocumentLevelMetricsByInstitutionId$endpoint
#> [1] "/DocumentLevelMetricsByInstitutionId/json"
#> 
#> 
#> $DocumentLevelMetricsByInstitutionIdRecordCount
#> $DocumentLevelMetricsByInstitutionIdRecordCount$description
#> [1] "Get metrics count by year"
#> 
#> $DocumentLevelMetricsByInstitutionIdRecordCount$verb
#> [1] "GET"
#> 
#> $DocumentLevelMetricsByInstitutionIdRecordCount$mime_type
#> [1] "json"
#> 
#> $DocumentLevelMetricsByInstitutionIdRecordCount$endpoint
#> [1] "/DocumentLevelMetricsByInstitutionIdRecordCount/json"
#> 
#> 
#> $DocumentLevelMetricsByUT
#> $DocumentLevelMetricsByUT$description
#> [1] "Get metrics by UT"
#> 
#> $DocumentLevelMetricsByUT$verb
#> [1] "GET"
#> 
#> $DocumentLevelMetricsByUT$mime_type
#> [1] "json"
#> 
#> $DocumentLevelMetricsByUT$endpoint
#> [1] "/DocumentLevelMetricsByUT/json"
#> 
#> 
#> $CheckValidUTs
#> $CheckValidUTs$description
#> [1] "Get invalid UTs"
#> 
#> $CheckValidUTs$verb
#> [1] "GET"
#> 
#> $CheckValidUTs$mime_type
#> [1] "json"
#> 
#> $CheckValidUTs$endpoint
#> [1] "/CheckValidUTs/json"
#> 
#> 
#> $InCitesLastUpdated
#> $InCitesLastUpdated$description
#> [1] "Get last updated information"
#> 
#> $InCitesLastUpdated$verb
#> [1] "GET"
#> 
#> $InCitesLastUpdated$mime_type
#> [1] "json"
#> 
#> $InCitesLastUpdated$endpoint
#> [1] "/InCitesLastUpdated/json"
#> 
#> 

ws_endpoints(api = c("incites", "endnote"))

#> $incites
#> $incites$DocumentLevelMetricsByInstitutionId
#> $incites$DocumentLevelMetricsByInstitutionId$description
#> [1] "Get metrics by institution Id"
#> 
#> $incites$DocumentLevelMetricsByInstitutionId$verb
#> [1] "GET"
#> 
#> $incites$DocumentLevelMetricsByInstitutionId$mime_type
#> [1] "json"
#> 
#> $incites$DocumentLevelMetricsByInstitutionId$endpoint
#> [1] "/DocumentLevelMetricsByInstitutionId/json"
#> 
#> 
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount$description
#> [1] "Get metrics count by year"
#> 
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount$verb
#> [1] "GET"
#> 
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount$mime_type
#> [1] "json"
#> 
#> $incites$DocumentLevelMetricsByInstitutionIdRecordCount$endpoint
#> [1] "/DocumentLevelMetricsByInstitutionIdRecordCount/json"
#> 
#> 
#> $incites$DocumentLevelMetricsByUT
#> $incites$DocumentLevelMetricsByUT$description
#> [1] "Get metrics by UT"
#> 
#> $incites$DocumentLevelMetricsByUT$verb
#> [1] "GET"
#> 
#> $incites$DocumentLevelMetricsByUT$mime_type
#> [1] "json"
#> 
#> $incites$DocumentLevelMetricsByUT$endpoint
#> [1] "/DocumentLevelMetricsByUT/json"
#> 
#> 
#> $incites$CheckValidUTs
#> $incites$CheckValidUTs$description
#> [1] "Get invalid UTs"
#> 
#> $incites$CheckValidUTs$verb
#> [1] "GET"
#> 
#> $incites$CheckValidUTs$mime_type
#> [1] "json"
#> 
#> $incites$CheckValidUTs$endpoint
#> [1] "/CheckValidUTs/json"
#> 
#> 
#> $incites$InCitesLastUpdated
#> $incites$InCitesLastUpdated$description
#> [1] "Get last updated information"
#> 
#> $incites$InCitesLastUpdated$verb
#> [1] "GET"
#> 
#> $incites$InCitesLastUpdated$mime_type
#> [1] "json"
#> 
#> $incites$InCitesLastUpdated$endpoint
#> [1] "/InCitesLastUpdated/json"
#> 
#> 
#> 
#> $<NA>
#> NULL
#> 

ws_incites_endpoints()

#> $DocumentLevelMetricsByInstitutionId
#> $DocumentLevelMetricsByInstitutionId$description
#> [1] "Get metrics by institution Id"
#> 
#> $DocumentLevelMetricsByInstitutionId$verb
#> [1] "GET"
#> 
#> $DocumentLevelMetricsByInstitutionId$mime_type
#> [1] "json"
#> 
#> $DocumentLevelMetricsByInstitutionId$endpoint
#> [1] "/DocumentLevelMetricsByInstitutionId/json"
#> 
#> 
#> $DocumentLevelMetricsByInstitutionIdRecordCount
#> $DocumentLevelMetricsByInstitutionIdRecordCount$description
#> [1] "Get metrics count by year"
#> 
#> $DocumentLevelMetricsByInstitutionIdRecordCount$verb
#> [1] "GET"
#> 
#> $DocumentLevelMetricsByInstitutionIdRecordCount$mime_type
#> [1] "json"
#> 
#> $DocumentLevelMetricsByInstitutionIdRecordCount$endpoint
#> [1] "/DocumentLevelMetricsByInstitutionIdRecordCount/json"
#> 
#> 
#> $DocumentLevelMetricsByUT
#> $DocumentLevelMetricsByUT$description
#> [1] "Get metrics by UT"
#> 
#> $DocumentLevelMetricsByUT$verb
#> [1] "GET"
#> 
#> $DocumentLevelMetricsByUT$mime_type
#> [1] "json"
#> 
#> $DocumentLevelMetricsByUT$endpoint
#> [1] "/DocumentLevelMetricsByUT/json"
#> 
#> 
#> $CheckValidUTs
#> $CheckValidUTs$description
#> [1] "Get invalid UTs"
#> 
#> $CheckValidUTs$verb
#> [1] "GET"
#> 
#> $CheckValidUTs$mime_type
#> [1] "json"
#> 
#> $CheckValidUTs$endpoint
#> [1] "/CheckValidUTs/json"
#> 
#> 
#> $InCitesLastUpdated
#> $InCitesLastUpdated$description
#> [1] "Get last updated information"
#> 
#> $InCitesLastUpdated$verb
#> [1] "GET"
#> 
#> $InCitesLastUpdated$mime_type
#> [1] "json"
#> 
#> $InCitesLastUpdated$endpoint
#> [1] "/InCitesLastUpdated/json"
#> 
#> 

ws_endnote_endpoints()

#> $user_space_attachment_size
#> $user_space_attachment_size$verb
#> [1] "GET"
#> 
#> $user_space_attachment_size$endpoint
#> [1] "/user_space/attachment_size"
#> 
#> $user_space_attachment_size$description
#> [1] "Retrieve the current total space of the all user's attachments"
#> 
#> 
#> $user_space_attachments
#> $user_space_attachments$verb
#> [1] "GET"
#> 
#> $user_space_attachments$endpoint
#> [1] "/user_space/attachments"
#> 
#> $user_space_attachments$description
#> [1] "Retrieve the total space available for a user"
#> 
#> 
#> $reference_types
#> $reference_types$verb
#> [1] "GET"
#> 
#> $reference_types$endpoint
#> [1] "/reference/types"
#> 
#> $reference_types$description
#> [1] "Retrieves the full collection of all reference types"
#> 
#> 
#> $user_space
#> $user_space$verb
#> [1] "GET"
#> 
#> $user_space$endpoint
#> [1] "/user_space"
#> 
#> $user_space$description
#> [1] "Retrieve user's space of the library"
#> 
#> 
#> $user_space_references
#> $user_space_references$verb
#> [1] "GET"
#> 
#> $user_space_references$endpoint
#> [1] "/user_space/references"
#> 
#> $user_space_references$description
#> [1] "Retrieve the amount of references that a user currently has, the amount of references a user can have in the library, and the amount of references that a user can still add."
#> 
#> 
#> $references_guid_file
#> $references_guid_file$verb
#> [1] "POST"
#> 
#> $references_guid_file$endpoint
#> [1] "/references/{guid}/file"
#> 
#> $references_guid_file$description
#> [1] "Create a file attached to this reference"
#> 
#> 
#> $reference_get
#> $reference_get$verb
#> [1] "GET"
#> 
#> $reference_get$endpoint
#> [1] "/reference"
#> 
#> $reference_get$description
#> [1] "Retrieve the reference specified in header x-enws-guid"
#> 
#> 
#> $reference_put
#> $reference_put$verb
#> [1] "PUT"
#> 
#> $reference_put$endpoint
#> [1] "/reference"
#> 
#> $reference_put$description
#> [1] "Change a set of a reference bibliographic content fields"
#> 
#> 
#> $reference_post
#> $reference_post$verb
#> [1] "POST"
#> 
#> $reference_post$endpoint
#> [1] "/reference"
#> 
#> $reference_post$description
#> [1] "Creates a reference from a set of its fields"
#> 
#> 
#> $references_guid_get
#> $references_guid_get$verb
#> [1] "GET"
#> 
#> $references_guid_get$endpoint
#> [1] "/references/{guid}"
#> 
#> $references_guid_get$description
#> [1] "Retreive the set of fields of a specific reference"
#> 
#> 
#> $references_guid_put
#> $references_guid_put$verb
#> [1] "PUT"
#> 
#> $references_guid_put$endpoint
#> [1] "/references/{guid}"
#> 
#> $references_guid_put$description
#> [1] "Change a set of this reference bibliographic content fields"
#> 
#> 
#> $root
#> $root$verb
#> [1] "GET"
#> 
#> $root$endpoint
#> [1] "/"
#> 
#> $root$description
#> [1] "Retrieve the user's Library object."
#> 
#> 

ws_woslite_endpoints()

#> $root
#> $root$verb
#> [1] "GET"
#> 
#> $root$endpoint
#> [1] "/"
#> 
#> $root$description
#> [1] "  Submits a user query and returns results"
#> 
#> 
#> $query
#> $query$verb
#> [1] "GET"
#> 
#> $query$endpoint
#> [1] "/query/{queryId}"
#> 
#> $query$description
#> [1] "Fetch record(s) by query identifier"
#> 
#> 
#> $id
#> $id$endpoint
#> [1] "/id/{uniqueId}"
#> 
#> $id$verb
#> [1] "GET"
#> 
#> $id$description
#> [1] "Find record(s) by specific id"
#> 
#> 

ws_woslite_endpoints()

#> $root
#> $root$verb
#> [1] "GET"
#> 
#> $root$endpoint
#> [1] "/"
#> 
#> $root$description
#> [1] "  Submits a user query and returns results"
#> 
#> 
#> $query
#> $query$verb
#> [1] "GET"
#> 
#> $query$endpoint
#> [1] "/query/{queryId}"
#> 
#> $query$description
#> [1] "Fetch record(s) by query identifier"
#> 
#> 
#> $id
#> $id$endpoint
#> [1] "/id/{uniqueId}"
#> 
#> $id$verb
#> [1] "GET"
#> 
#> $id$description
#> [1] "Find record(s) by specific id"
#> 
#> 

ws_tipms_endpoints()

#> $case_record_post
#> $case_record_post$verb
#> [1] "POST"
#> 
#> $case_record_post$endpoint
#> [1] "/CaseRecord/{ipType}"
#> 
#> $case_record_post$description
#> [1] "Posts (Creates) the fields for case record."
#> 
#> 
#> $case_record_put
#> $case_record_put$verb
#> [1] "PUT"
#> 
#> $case_record_put$endpoint
#> [1] "/CaseRecord/{ipType}/{masterId}"
#> 
#> $case_record_put$description
#> [1] "Puts (Updates) the fields for specified case record."
#> 
#> 
#> $file
#> $file$verb
#> [1] "POST"
#> 
#> $file$endpoint
#> [1] "/File"
#> 
#> $file$description
#> [1] "Posts selected file."
#> 
#> 
#> $health
#> $health$verb
#> [1] "GET"
#> 
#> $health$endpoint
#> [1] "/Health"
#> 
#> $health$description
#> [1] "Reflects the state of the application."
#> 
#> 
#> $patent_filing_forms
#> $patent_filing_forms$verb
#> [1] "GET"
#> 
#> $patent_filing_forms$endpoint
#> [1] "/patentfilingforms"
#> 
#> $patent_filing_forms$description
#> [1] "Gets (Retrieves) the patent filing forms data."
#> 
#> 
#> $tempfile
#> $tempfile$verb
#> [1] "POST"
#> 
#> $tempfile$endpoint
#> [1] "/TempFile"
#> 
#> $tempfile$description
#> [1] "Posts (Creates) the selected file."
#> 
#> 
#> $trademark_reporting
#> $trademark_reporting$verb
#> [1] "GET"
#> 
#> $trademark_reporting$endpoint
#> [1] "/trademarkreporting"
#> 
#> $trademark_reporting$description
#> [1] "Search for Trademarks reporting data."
#> 
#>